+++ TestApp/AppDelegate.swift
@@ -2,19 +2,22 @@
 //  AppDelegate.swift
 //  TestApp
 //
-//  Created by Antihevich on 8/5/16.
-//  Copyright © 2016 Artem Antihevich. All rights reserved.
+//  Created by Artem Antihevich on 7/26/18.
+//  Copyright © 2018 Artem Antihevich. All rights reserved.
 //
 
 import UIKit
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
+
     var window: UIWindow?
+
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
         return true
     }
+
 }